fix(MaterialColor): Removed circular dependency warnings
BREAKING CHANGE: Static method MaterialColor.fromText is replaced with (new MaterialColors()).getColorFromText
