+++ src/common/core/style/material-color.ts
@@ -7,19 +7,8 @@
  */
 
 import { Color } from '../graphics/color';
-import { MaterialColors } from './material-colors';
 
 export class MaterialColor extends Color {
     public name: string;
     public subColors = [] as MaterialColor[];
-
-    public static fromText(text: string) {
-        let sum = 0;
-        for (let i = 0; i < text.length; i++) {
-            sum += text.charCodeAt(i);
-        }
-        const colors = new MaterialColors().colors;
-        const subColors = (colors[sum % colors.length]).subColors;
-        return subColors[sum % subColors.length];
-    }
 }
